change inotify

+++ __init__.py
@@ -18,18 +18,17 @@ class EntryEventHandler(pyinotify.ProcessEvent):
     """
     EntryEventHandler monitor entries added, modified or deleted
     """
-    def process_IN_CREATE(self, event):
-        entryService.add_entry(True, event.pathname)
-
-    def process_IN_MODIFY(self, event):
-        entryService.add_entry(True, event.pathname)
-        
-    def process_IN_DELETE(self, event):
-        entryService.delete_entry(event.pathname)
+    def process_default(self, event):
+        mask_add = pyinotify.IN_CLOSE_WRITE | pyinotify.IN_MOVED_TO | pyinotify.IN_MOVE_SELF
+        mask_del = pyinotify.IN_DELETE | pyinotify.IN_DELETE_SELF | pyinotify.IN_MOVED_FROM | pyinotify.IN_MOVE_SELF
+        if event.mask & mask_add:
+            entryService.add_entry(True, event.pathname)
+        if event.mask & mask_del:
+            entryService.delete_entry(event.pathname)
 
 
 wm = pyinotify.WatchManager()
-mask = pyinotify.IN_CREATE | pyinotify.IN_MODIFY | pyinotify.IN_DELETE
+mask = pyinotify.ALL_EVENTS
 path =  config.entry_dir
 notifier = pyinotify.ThreadedNotifier(wm, EntryEventHandler())
 wdd = wm.add_watch(path, mask, rec=True)
